Configuration/travis (#4)
* Modification of TravisFile

We still have some PHPunit testing pending.

* autoload file

We included an autoload file.

* Travis with Symfony version

We pay attention to the version of PHP for Symphony 3.*

* Testing with Travis

* More Testing Travis

Adding Symfony 3.3
Verify PHP 5.4.*

* Configuration Travis et Testing

The testing has been properly fixed.

The configuration for Travis has been modified.

* Configuration Travis CI

We made other modification. We added the notification email and the
phpunit script.

Symfony 2.8.x only on PHP 5.x

* Added phpunit

* Travis

* Travis

* Travis

* Change composer, gitignore, autoload

The autoload has been modified.
The composer.json has been adjusted for the Testing.

* Testing only the bundle.

* Symfony 3.1 will not be supporting in July 2017

Symfony 3.1 will not be supporting in July 2017. It will be removed in a
next version.

* Travis minus Symfony 3.1